Rename `CodeLayout` to `SourceLayout`.
